refactor shop page component layout
