+++ src/components/HomePageProductCardHeader.jsx
@@ -1,0 +1,13 @@
+export default function HomePageProductCardHeader({ product }) {
+  return (
+    <div>
+      <div className="mb-8 text-center">
+        <h2 className="text-lg text-gray-500">Featured Products</h2>
+        <h1 className="text-xl font-semibold uppercase">Bestseller Products</h1>
+        <p className="text-gray-400">
+          Problems trying to resolve the conflict between
+        </p>
+      </div>
+    </div>
+  );
+}
