+++ src/components/ShopPageFilter.jsx
@@ -1,0 +1,3 @@
+export default function ShopPageFilter() {
+  return <div>Filter</div>;
+}
