+++ src/components/ShopPageHeader.jsx
@@ -1,0 +1,3 @@
+export default function ShopPageHeader() {
+  return <div>Header</div>;
+}
